Addressing CR comments and reformatting code to align with Code-Style
