+++ xchange-cexio/src/main/java/org/knowm/xchange/cexio/service/CexIOMarketDataServiceRaw.java
@@ -1,5 +1,7 @@
 package org.knowm.xchange.cexio.service;
 
+import java.io.IOException;
+
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.cexio.CexIO;
 import org.knowm.xchange.cexio.dto.marketdata.CexIOCurrencyLimits;
@@ -7,9 +9,8 @@ import org.knowm.xchange.cexio.dto.marketdata.CexIODepth;
 import org.knowm.xchange.cexio.dto.marketdata.CexIOTicker;
 import org.knowm.xchange.cexio.dto.marketdata.CexIOTrade;
 import org.knowm.xchange.currency.CurrencyPair;
-import si.mazi.rescu.RestProxyFactory;
 
-import java.io.IOException;
+import si.mazi.rescu.RestProxyFactory;
 
 /**
  * @author timmolter
