+++ xchange-cexio/src/test/java/org/knowm/xchange/cexio/service/marketdata/RemoteInitTest.java
@@ -1,17 +1,14 @@
 package org.knowm.xchange.cexio.service.marketdata;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.Test;
-import org.knowm.xchange.Exchange;
-import org.knowm.xchange.ExchangeFactory;
-import org.knowm.xchange.cexio.CexIOExchange;
-import org.knowm.xchange.cexio.dto.marketdata.CexIOCurrencyLimits;
-import org.knowm.xchange.cexio.service.CexIOMarketDataService;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
 import java.io.InputStream;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import org.junit.Test;
+import org.knowm.xchange.cexio.dto.marketdata.CexIOCurrencyLimits;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
  * @author ujjwal on 13/02/18.
@@ -33,17 +30,4 @@ public class RemoteInitTest {
 
   }
 
-  @Test
-  public void integrationTest() throws IOException {
-    final Exchange exchange = ExchangeFactory.INSTANCE.createExchange(CexIOExchange.class);
-
-    assertThat(exchange).isNotNull();
-    assertThat(exchange.getExchangeMetaData().getCurrencyPairs()).isNotEmpty();
-
-    final CexIOCurrencyLimits currencyLimits = ((CexIOMarketDataService) exchange.getMarketDataService()).getCurrencyLimits();
-    assertThat(currencyLimits.getE()).isEqualToIgnoringCase("currency_limits");
-    assertThat(currencyLimits.getError()).isNullOrEmpty();
-    assertThat(currencyLimits.getOk()).isEqualToIgnoringCase("ok");
-
-  }
 }
